% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_aliases.R
\name{klue-deprecated}
\alias{klue-deprecated}
\alias{run_lcmnl_workflow}
\alias{build_database}
\alias{build_database_long}
\alias{build_database_wide}
\alias{make_dgp_config}
\alias{estimate_lcmnl_multistart}
\alias{estimate_mmnl}
\alias{estimate_mmnl_corr}
\alias{estimate_lcmnl_multistart_onehot}
\alias{generate_data}
\alias{generate_data_with_covariates}
\alias{generate_data_defficient}
\alias{run_full_study}
\alias{run_main_simulation}
\alias{run_mmnl_comparison}
\alias{run_convergence_ablation}
\alias{run_initialisation_ablation}
\alias{run_unbalanced_analysis}
\alias{run_design_comparison}
\alias{run_concomitant_analysis}
\alias{run_unconditional_recovery}
\alias{run_clustering_comparison}
\alias{run_sample_sensitivity}
\alias{run_correlated_mmnl_robustness}
\title{Deprecated backward-compatibility aliases}
\usage{
run_lcmnl_workflow(...)

build_database(...)

build_database_long(...)

build_database_wide(...)

make_dgp_config(...)

estimate_lcmnl_multistart(...)

estimate_mmnl(...)

estimate_mmnl_corr(...)

estimate_lcmnl_multistart_onehot(database, C, dgp = DGP_DEFAULT)

generate_data(...)

generate_data_with_covariates(...)

generate_data_defficient(...)

run_full_study(...)

run_main_simulation(...)

run_mmnl_comparison(...)

run_convergence_ablation(...)

run_initialisation_ablation(...)

run_unbalanced_analysis(...)

run_design_comparison(...)

run_concomitant_analysis(...)

run_unconditional_recovery(...)

run_clustering_comparison(...)

run_sample_sensitivity(...)

run_correlated_mmnl_robustness(...)
}
\description{
Each function listed here forwards to its canonical \code{klue_} function.
These aliases are kept so older scripts that use the previous names keep
working.
}
\keyword{internal}
